Make SkPictures valid regardless of SK_SUPPORT_HINTING_SCALE_FACTOR.
When the build flag is not set, read/write dummy values
so that the creator of an SKP file need not have the same
support/lack of support of the feature as the reader.

Will separately update the checked in skps to the new version when checking in.

BUG=http://code.google.com/p/skia/issues/detail?id=922

Review URL: https://codereview.appspot.com/6642057

git-svn-id: http://skia.googlecode.com/svn/trunk/src@5869 2bbb7eff-a529-9590-31e7-b0007b416f81