Only print error message in debug mode.
Currently, PictureTest runs a test of recording a bad bitmap to a
picture and then replaying the picture. We assert at the record
step in debug mode, so the test only runs in release, to ensure that
we can still play back a picture, even if it has a bad bitmap. Silence
the error message in release, so our test doesn't appear to error
(since it is actually working as desired).

Review URL: https://codereview.chromium.org/12315138

git-svn-id: http://skia.googlecode.com/svn/trunk/src@7897 2bbb7eff-a529-9590-31e7-b0007b416f81