Shader optimization for ovals.
Previously we were setting a center point as an attribute 
across all vertices, then using that with the interpolated 
position to compute an offset. However, because the offset 
computation is linear, we can just set the appropriate offset 
at the vertices and let the rasterizer interpolate it for us. 

Author: jvanverth@google.com

Reviewed By: robertphillips@google.com

Review URL: https://chromiumcodereview.appspot.com/14093002

git-svn-id: http://skia.googlecode.com/svn/trunk/src@8613 2bbb7eff-a529-9590-31e7-b0007b416f81