Support for libpng > 1.2.
  * Add conditional defines for deprecated NULL type casts.
  * Use wrapper functions for accessing png_structp & png_infop fields
    (direct access is deprecated).

Tested on Fedora 17 (libpng 1.5) and Ubuntu 12.04 (libpng 1.2).

Note: there is one png_structp direct field acces left in
SkPNGImageDecoder::onDecodeRegion() for which libpng does not provide a
wrapper method (png_ptr->pass = 0). That should be OK for now as the code
is only enabled when SK_BUILD_FOR_ANDROID is set, and the Android libpng
version is striclty controlled. But it may become an issue in the future
when/if Android decides to update libpng, so we should investigate a
solution.


R=scroggo@google.com, reed@google.com, djsollen@google.com

Author: fmalita@google.com

Review URL: https://chromiumcodereview.appspot.com/14645018

git-svn-id: http://skia.googlecode.com/svn/trunk/src@8956 2bbb7eff-a529-9590-31e7-b0007b416f81