The CL adds libpoppler to DEPS and adds a libpoppler-cpp gyp target for Linux, Windows, and Mac. This does not currently change the GM tool to use poppler - that will be a refactor job in a separate CL.
Several third-party libraries are pulled in, and platform-specific headers are included under third_party/(library name). Chromium style READMEs containing a description of the library, as well as a LICENSE file are also included in that directory. (is there a Skia-specific style for these?)

R=vandebo@chromium.org, edisonn@google.com, djsollen@chromium.org, bungeman@chromium.org, open-source-third-party-reviews@google.com, thestig@chromium.org

Author: richardlin@chromium.org

Review URL: https://chromiumcodereview.appspot.com/20220002

git-svn-id: http://skia.googlecode.com/svn/trunk/src@10823 2bbb7eff-a529-9590-31e7-b0007b416f81