An SkOSWindow for Unix may experience stuttering, with missed events or long delays before events are processed when using SkEvent.postDelay().
The issue is with the use of a select loop to watch for events coming into the x11 file descriptor for the display. It turns out there may be XEvents queued in memory but not yet processed that need to be handled before entering the select loop. See: http://developerweb.net/viewtopic.php?id=3184

BUG=1960
R=reed@google.com

Author: jcgregorio@google.com

Review URL: https://codereview.chromium.org/123303004

git-svn-id: http://skia.googlecode.com/svn/trunk/src@12874 2bbb7eff-a529-9590-31e7-b0007b416f81